% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_meta_analysis.R
\name{bma}
\alias{bma}
\title{Bayesian Meta-analysis}
\usage{
bma(data, plot = TRUE, model = "random_walk", parallel = FALSE,
  incl.model = TRUE, n.iter = 10000, m.scale = "loge")
}
\arguments{
\item{data}{a data.frame with 4 columns in this order: species, year, index, se (standard error)}

\item{plot}{Logical, should a trace plot be plotted?}

\item{model}{The type of model to be used. The default of 'random_walk' is strongly recommended.
Two other models 'uniform' and 'uniform_noeta' exist but are almost always inferior.}

\item{parallel}{if \code{TRUE} the model chains will be run in parallel using one fewer cores than
are availbale on the computer.}

\item{incl.model}{if \code{TRUE} the model is added as an attribute of the object returned}

\item{n.iter}{The number of iterations of the model to run. Defaults to 10,000 to avoid long run times
though much longer runs are usually required to reach convergence}

\item{m.scale}{The measurement scale of the data. The scale of the data is assumed to be logarithmic.
Here you specify which log scale the data is on ('loge', 'log10', or 'logit'). Defaults to 'loge'.}
}
\value{
Returns a dataframe with 4 columns: Year, Index, lower2.5, upper97.5. The last two columns are the credible intervals
}
\description{
Use a bayesian meta-analysis to create an indicator from species index values and standard error.
}
\examples{

# Create some example data in the format required
data <- data.frame(species = rep(letters, each = 50),
                   year = rep(1:50, length(letters)),
                   index = runif(n = 50 * length(letters), min = 0, max = 1),
                   se = runif(n = 50 * length(letters), min = 0.01, max = .1))

# Run the Bayesian meta-analysis
bma_indicator <- bma(data)

# Plot the resulting indicator
plot_indicator(indicator = bma_indicator[,'Index'],
               CIs = bma_indicator[,c(3,4)])
}

