% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap_indicator.R
\name{bootstrap_indicator}
\alias{bootstrap_indicator}
\title{Bootstrapping indicator}
\usage{
bootstrap_indicator(Data, iterations = 10000, CI_limits = c(0.025, 0.975),
  verbose = TRUE)
}
\arguments{
\item{Data}{A matrix, the first column gives the year, subsequent named
columns give the species and the named column 'indicator' gives
the indicator value.}

\item{iterations}{The number of bootstrap iterations to use.}

\item{CI_limits}{The confidence limits to return as a vector of two
numbers. This default to c(0.025, 0.975) the 95 percent conficence intervals.}

\item{verbose}{If \code{TRUE} then progress is printed to screen.}
}
\value{
A matrix. In each row we have the year, the species and the
        scaled value. There is also an additional column, 'geomean' giving
        the geometric mean for each year.
}
\description{
This function takes in a dataframe of multi-species data that have been scaled
and adds confidence intervals to the indicator value.
}

