% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_jags.r
\name{detect_jags}
\alias{detect_jags}
\title{Detect JAGS installation}
\usage{
detect_jags()
}
\value{
TRUE or FALSE, indicating whether the JAGS installation has been detected
}
\description{
small function that detects whether JAGS is installed.
}
\examples{

detect_jags()
}
