% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM_smoothing.R
\name{GAM_smoothing}
\alias{GAM_smoothing}
\title{GAM smoothing}
\usage{
GAM_smoothing(values, years = 1:length(values), plot = FALSE, ...)
}
\arguments{
\item{values}{A numeric vector that will be smoothed}

\item{years}{Optional, the years that \code{value} corresponds
to as a numeric vector.}

\item{plot}{If \code{TRUE} a plot is generated showing the smoothed
spline.}

\item{...}{other arguements passed to \code{mgcv::gam}}
}
\value{
The smoothed/predicted values are returned as well as the 
model object used for making predictions.
}
\description{
This function uses a GAM to smooth a numeric vector
}
