% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bmaBUGScode.R
\name{get_bmaBUGScode}
\alias{get_bmaBUGScode}
\title{Extract BMA BUGS code}
\usage{
get_bmaBUGScode(option = "Smooth", print.screen = FALSE,
  save.local = FALSE)
}
\arguments{
\item{option}{text string defining which particular variant is desired}
}
\description{
Gets a copy of the BUGS code and writes it
}
\examples{
extract_bmaBUGScode(option="Smooth", print.screen=TRUE, save.local=FALSE)
}
