% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambda_interpolation.r
\name{lambda_interpolation}
\alias{lambda_interpolation}
\title{Rescale species values for indicator using Lambda interpolation}
\usage{
lambda_interpolation(input, index = 100, threshold_sd = 0.2,
  threshold_yrs = 20, upperQuantile = 0.975, lowerQuantile = 0.025,
  logOdds = TRUE)
}
\arguments{
\item{input}{Either a string giving the path to occupancy model output files
produced by sparta, or a three dimensional array [species, year, iteration].}

\item{index}{The index value for the first year, defaults to 100.}

\item{threshold_sd}{The threshold standard deviation for a species-year value
to be included. If the standard deviation is above this value it is removed.}

\item{threshold_yrs}{Numeric, the minimum number of years that a species must
fulfill the threshold_sd for it to be included.}

\item{upperQuantile}{The upper confidence interval to use (as a probability)}

\item{lowerQuantile}{The lower confidence interval to use (as a probability)}

\item{logOdds}{Logical, if \code{TRUE} then log odds}
}
\value{
A list with four elements: a summary (data.frame), the raw data used
in the indicator (after removing species that fail thresholds), the raw
indicator value (a value for each iteration in each year), and a measure of
change for each species [ADD SPECIFICS].
}
\description{
This function takes in the output from a sparta occupancy model
or a three dimensional array. The first year is set to an index value for all
species and then using the average change for one year to the next, across
species is used to calculate the indicator. A species only contributes to the
dataset in years where the standard deviation across iterations
meets the sd threshold. Missing years in the middle a a species dataset are
filled in using interpolation.
}
\examples{
### Running from an array ####
# number of species
nsp = 50

# number of years
nyr = 40

#number of iterations
iter = 500

# Build a random set of data
myArray <- array(rnorm(n = nsp*nyr*iter,
                 mean = 0.5,
                 sd = 0.1),
                 c(nsp, nyr, iter))

# Ensure values are bounded by 0 and 1
myArray[myArray > 1] <- 1
myArray[myArray > 0] <- 0

# Run the lambda_interpolation method on this data
myIndicator <- lambda_interpolation(myArray)

# Plot the indicator
plot_indicator(myIndicator$summary[,'indicator'],
               myIndicator$summary[,c('2.5\%' ,'97.5\%')])

### Running from a directory of sparta ouput
# myIndicator <- lambda_interpolation('myfilepath/directory')#'
}

