% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MSI.R
\name{plot.MSI}
\alias{plot.MSI}
\title{A function to plots msi results}
\usage{
\method{plot}{MSI}(x, title = "MSI results", ...)
}
\arguments{
\item{x}{object of class MSI}

\item{title}{Character, the title for the plot}

\item{\dots}{currently ignored}
}
\description{
Plots a summary of the MSI analysis
}

