% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_stack.r
\name{plot_trend_stack}
\alias{plot_trend_stack}
\title{Plot species trend stack plots}
\usage{
plot_trend_stack(species_change)
}
\arguments{
\item{species_change}{factor vector of change categories. This should have the
following levels "strong increase", "increase", "no change", "decrease",
"strong decrease".}
}
\value{
Returns a data.frame of the data in the plot. A plot is sent to the device.
}
\description{
This function is used for reporting on species trends generated
in indicators. It takes a vector of categories and plots then in a stacked
column plot
}
\examples{
### Running from an array ####
set.seed(123)
# number of species
nsp = 50

# number of years
nyr = 40

#number of iterations
iter = 3

# Build a random set of data
myArray <- array(rnorm(n = nsp*nyr*iter,
                 mean = 0.5,
                 sd = 0.2),
                 c(nsp, nyr, iter))

# Ensure values are bounded by 0 and 1
myArray[myArray > 1] <- 1
myArray[myArray < 0] <- 0

# Run the lambda_indicator method on this data                
myIndicator <- lambda_indicator(myArray)

# Plot the trend stack
plot_trend_stack(myIndicator$species_change[,'category'])
}
