% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rescale_posterior.r
\name{rescale_posterior}
\alias{rescale_posterior}
\title{Rescale species values for indicator}
\usage{
rescale_posterior(input_dir, subset_table = NULL, index = 100,
  max = 10000, min = 1, year_limit = 10, upperCI = 0.975,
  lowerCI = 0.025, verbose = TRUE)
}
\arguments{
\item{input_dir}{the loction of occupancy model output files}

\item{subset_table}{dataframe with columns for years and rows for species.
Column names must be character years and row names must match the file names
in input_dir}

\item{index}{The index value for the first year, defaults to 100.}

\item{max}{The upper limit allowed for scaled values. Values greater than
this will be set equal to \code{max}.}

\item{min}{The upper limit allowed for scaled values. Values greater than
this will be set equal to \code{min}.}

\item{year_limit}{The minimum length of 'good years' for a species to be
included}

\item{upperCI}{The upper confidence interval to use (as a probability)}

\item{lowerCI}{The lower confidence interval to use (as a probability)}

\item{verbose}{If TRUE progress is written to console}
}
\value{
A list with two elements, a summary and the rescaled data
}
\description{
This function takes in a dataframe of multi-species data and rescales them
so that the value in the starting year is the same. This function accounts
for species that have no data at the beginning of the period or those that
have no data at teh end of the period.
}

