% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_occDet.R
\name{summarise_occDet}
\alias{summarise_occDet}
\title{Summarise Occupancy model output}
\usage{
summarise_occDet(input_dir, region = NULL, verbose = TRUE)
}
\arguments{
\item{input_dir}{The directory where all the .rdata files are located.
This will be the same as the \code{output_dir} arguement given
to occDetModel in sparta when creating the output.}

\item{region}{If the model output contains regional estimates specify the region
name here to extract estimates from that region only (e.g."ENGLAND"). If NULL
the estimate based on all sites will be used instead.}

\item{verbose}{If \code{TRUE} then progress is printed to screen.}
}
\value{
A data.frame. In each row we have the year, the species and the
        mean of the posterior for the predicted proportion of sites occupied.
}
\description{
This function reads in and summarises all the .rdata files that are output
by the occDetModel in sparta.
}
