% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bats-data.R
\docType{data}
\name{bats}
\alias{bats}
\title{Data - National Bat Monitoring Programme, UK}
\format{An object of class \code{data.frame} with 136 rows and 3 columns.}
\source{
\url{https://doi.org/10.1016/j.biocon.2014.11.022}
}
\usage{
data(bats)
}
\description{
This dataset from the National Bat Monitoring Programme, UK has data for UK bats species
The data contains: Species, an abbreviation of the bat's name; Year, integer
from 1998-2014; Index, the index values are unsmoothed, log (base e)
transformations of projected counts. For more information see Barlow et al
}
\examples{
data(bats)
head(bats)
}
\references{
Barlow, K. E., Briggs, P. A., Haysom, K. A., Hutson, A. M., Lechiara, N. L., Racey, P. A., … Langton, S. D. (2015). Citizen science reveals trends in bat populations: The National Bat Monitoring Programme in Great Britain. Biological Conservation, 182, 14–26. \url{https://doi.org/10.1016/j.biocon.2014.11.022}
}
\keyword{datasets}
