% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_meta_analysis.R
\name{bma}
\alias{bma}
\title{Bayesian Meta-analysis}
\usage{
bma(data, randwalk = TRUE, plot = TRUE)
}
\arguments{
\item{data}{a data.frame with 4 columns in this order: species, year, index, se (standard error)}

\item{randwalk}{If \code{TRUE} a random walk feature is added to the model (defualt), else years are treated as independent.}

\item{plot}{Logical, should a trace plot be plotted?}
}
\value{
Returns a dataframe with 4 columns: Year, Index, lower2.5, upper97.5. The last two columns are the credible intervals
}
\description{
Use a bayesian meta-analysis to create an indicator from species index values and standard error.
}
\examples{

# Create some example data in the format required
data <- data.frame(species = rep(letters, each = 50),
                   year = rep(1:50, length(letters)),
                   index = runif(n = 50 * length(letters), min = 0, max = 1),
                   se = runif(n = 50 * length(letters), min = 0.01, max = .1))

# Run the Bayesian meta-analysis
bma_indicator <- bma(data)

# Plot the resulting indicator
plot_indicator(indicator = bma_indicator[,'Index'],
               CIs = bma_indicator[,c(3,4)])
}

