% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_meta_analysis.R
\name{bma}
\alias{bma}
\title{Bayesian Meta-analysis}
\usage{
bma(
  data,
  plot = TRUE,
  model = "smooth",
  parallel = FALSE,
  incl.model = TRUE,
  n.iter = 10000,
  m.scale = "loge",
  num.knots = 12,
  rescaleYr = 1,
  seFromData = FALSE,
  Y1perfect = TRUE,
  incl.2deriv = FALSE,
  save.sppars = TRUE,
  n.thin = 5,
  q = c(0.025, 0.975)
)
}
\arguments{
\item{data}{a data.frame with 3-4 columns in this order: `species`, `year`, `index`, `se` (standard error). The `se` column is optional 
NB: Index values are assumed to be on the unbounded (logarithmic scale)}

\item{plot}{Logical, should a trace plot be plotted to diagnose the model output?}

\item{model}{The type of model to be used. See details.}

\item{parallel}{if \code{TRUE} the model chains will be run in parallel using one fewer cores than
are available on the computer. NOTE: this will typically not work for parallel use on cluster PCs.}

\item{incl.model}{if \code{TRUE} the model is added as an attribute of the object returned}

\item{n.iter}{The number of iterations of the model to run. Defaults to 10,000 to avoid long run times
though much longer runs are usually required to reach convergence}

\item{m.scale}{The measurement scale of the data. The scale of the data is assumed to be logarithmic.
Here you specify which log scale the data is on ('loge', 'log10', or 'logit'). Defaults to 'loge'.}

\item{num.knots}{If using either of the smooth models this specifies the number of knots.}

\item{rescaleYr}{Logical, should all iterations be scaled so that the first year is equal? If TRUE
year one will have 0 error.}

\item{seFromData}{Logical. Should the standard errors be read in from data (`TRUE`) or estimated (`FALSE`)?  Defaults to `FALSE`}

\item{Y1perfect}{Logical. Should the first year of a species' index be assumed known without error (`TRUE`)? Defaults to `TRUE`}

\item{incl.2deriv}{Logical. Option to include estimation of second derivatives on the indicator (`TRUE`)? Defaults to `FALSE`}

\item{save.sppars}{Logical. Should the species-specific parameters be monitored? Defaults to TRUE}

\item{n.thin}{Thinning rate for the Markov chains. Defaults to 5.}

\item{q}{defines the quantiles of the posterior distribution to report. Defaults to c(0.025, 0.975), i.e. the 95th percentile credible intervals}
}
\value{
Returns a dataframe with 4 columns: Year, Index, lower2.5, upper97.5. The last two columns are the credible intervals
}
\description{
Use a Bayesian meta-analysis to create an indicator from species index values, optionally incorporating standard error.
}
\details{
There are a number of model to choose from:
\itemize{
 \item{\code{"smooth"}}{ The default option. Indicator defined by Growth rates, with Ruppert smoother, allowing for species to join late. Error on the first year of each species' time-series is assumed to be zero. The indicator is the expected value of the geometric mean across species (with missing species imputed). 
 Includes three options: `seFromData` `Y1perfect` and `incl.2deriv`. See bayesian_meta_analysis for mode details. Using the default values `seFromData = FALSE` and `Y1perfect = TRUE` are the options used in Freeman  \emph{et al.} (2020).}
 \item{\code{"smooth_det2"}}{ Equivalent to smooth with `seFromData = TRUE` and `Y1perfect = FALSE`. Retained for backwards compatability. Choosing this option will overwrite user-entered options for `seFromData` and `Y1perfect`.}
 \item{\code{"smooth_det_sigtheta"}}{ Equivalent to smooth with `seFromData = FALSE` and `Y1perfect = FALSE`. Retained for backwards compatability. Choosing this option will overwrite user-entered options for `seFromData` and `Y1perfect`.}
 \item{\code{"smooth_det"}}{ Specific variant of smooth_det2 - under review. Likely to be deprecated}
 }
}
\examples{
# Create some example data in the format required
data <- data.frame(species = rep(letters, each = 50),
                   year = rep(1:50, length(letters)),
                   index = rnorm(n = 50 * length(letters), mean = 0, sd = 1),
                   se = runif(n = 50 * length(letters), min = 0.01, max = .1))

# Run the Bayesian meta-analysis
bma_indicator <- bma(data, model="smooth", m.scale="logit", n.iter=100)

# Plot the resulting indicator
plot_indicator(indicator = bma_indicator[,'Index'],
               CIs = bma_indicator[,c(3,4)])
}
\references{
Freeman, S.N., Isaac, N.J.B., Besbeas, P.T., Dennis, E.B. & Morgan, B.J.T. (2019) 
            A generic method for estimating and smoothing multispecies biodiversity indices, robust to intermittent data. 
            \emph{JABES}, in revision.
}
