% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_meta_analysis.R
\name{bma}
\alias{bma}
\title{Bayesian Meta-analysis}
\usage{
bma(data, plot = TRUE, model = "random_walk", parallel = FALSE,
  incl.model = TRUE, n.iter = 10000, m.scale = "loge",
  num.knots = 12, rescaleYr = 1, n.thin = 5, save.sppars = TRUE)
}
\arguments{
\item{data}{a data.frame with 4 columns in this order: species, year, index, se (standard error)}

\item{plot}{Logical, should a trace plot be plotted?}

\item{model}{The type of model to be used. See details.}

\item{parallel}{if \code{TRUE} the model chains will be run in parallel using one fewer cores than
are available on the computer. NOTE: this will typically not work for parallel use on cluster PCs.}

\item{incl.model}{if \code{TRUE} the model is added as an attribute of the object returned}

\item{n.iter}{The number of iterations of the model to run. Defaults to 10,000 to avoid long run times
though much longer runs are usually required to reach convergence}

\item{m.scale}{The measurement scale of the data. The scale of the data is assumed to be logarithmic.
Here you specify which log scale the data is on ('loge', 'log10', or 'logit'). Defaults to 'loge'.}

\item{num.knots}{If using either of the smooth models this specifies the number of knots.}

\item{n.thin}{Thinning rate for the Markov chains. Defaults to 5.}

\item{save.sppars}{Logical. Should the species-specific parameters be monitored? Defaults to TRUE}

\item{rescaleYr1}{Logical, should all iterations be scaled so that the first year is equal? If TRUE
year one will have 0 error.}
}
\value{
Returns a dataframe with 4 columns: Year, Index, lower2.5, upper97.5. The last two columns are the credible intervals
}
\description{
Use a bayesian meta-analysis to create an indicator from species index values and standard error.
}
\details{
There are a number of model to choose from:
\itemize{
 \item{\code{"random_walk"}}{ - Also known as BMA3.}
 \item{\code{"uniform"}}{ - Also known as BMA2.}
 \item{\code{"uniform_noeta"}}{ - Also known as BMA1.}
 \item{\code{"FNgr"}}{ - Indicator defined by Growth rates, as in Freeman & Newson.}
 \item{\code{"smooth_stoch"}}{ - Indicator defined by Growth rates, with Ruppert smoother (stochastic version).}
 \item{\code{"smooth_det"}}{ - Indicator defined by Growth rates, with Ruppert smoother (deterministic version).}
 \item{\code{"FNgr2"}}{ - Variant where species can join the series late and error on the first year is 0 (check with Nick and Steve).}
 \item{\code{"smooth_stoch2"}}{ - Variant where species can join the series late and error on the first year is 0 (check with Nick and Steve).}
 \item{\code{"smooth_det2"}}{ - Variant where species can join the series late and error on the first year is 0 (check with Nick and Steve).}
 \item{\code{"smooth_det_sigtheta"}}{ - Variant of det2 in which standard errors are assumed constant (check with Nick and Steve).}
}
}
\examples{

# Create some example data in the format required
data <- data.frame(species = rep(letters, each = 50),
                   year = rep(1:50, length(letters)),
                   index = runif(n = 50 * length(letters), min = 0, max = 1),
                   se = runif(n = 50 * length(letters), min = 0.01, max = .1))

# Run the Bayesian meta-analysis
bma_indicator <- bma(data)

# Plot the resulting indicator
plot_indicator(indicator = bma_indicator[,'Index'],
               CIs = bma_indicator[,c(3,4)])
}
