% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCindicators-package.R
\docType{package}
\name{BRCindicators-package}
\alias{BRCindicators}
\alias{BRCindicators-package}
\title{BRCindicators: Creating multispecies biodiversity indicators}
\description{
This package contains a number of functions used to create biodiversity indicators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BiologicalRecordsCentre/BRCindicators}
  \item Report bugs at \url{https://github.com/BiologicalRecordsCentre/BRCindicators/issues}
}

}
\author{
\strong{Maintainer}: Tom August \email{tomaug@ceh.ac.uk} (\href{https://orcid.org/0000-0003-1116-3385}{ORCID})

Authors:
\itemize{
  \item Gary Powney (\href{https://orcid.org/0000-0003-3313-7786}{ORCID})
  \item Charlie Outhwaite (\href{https://orcid.org/0000-0001-9997-6780}{ORCID})
  \item Jack Hatfield (\href{https://orcid.org/0000-0002-6361-0629}{ORCID})
  \item Mark Logie (\href{https://orcid.org/0000-0003-0840-9575}{ORCID})
  \item Stephen Freeman
  \item Nick Isaac (\href{https://orcid.org/0000-0002-4869-8052}{ORCID})
}

}
\keyword{internal}
