% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCindicators-package.R
\docType{package}
\name{BRCindicators-package}
\alias{BRCindicators}
\alias{BRCindicators-package}
\title{BRCindicators: Creating multispecies biodiversity indicators}
\description{
This package contains a number of functions used to create biodiversity indicators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BiologicalRecordsCentre/BRCindicators}
  \item Report bugs at \url{https://github.com/BiologicalRecordsCentre/BRCindicators/issues}
}

}
\keyword{internal}
