% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bmaBUGScode.R
\name{get_bmaBUGScode}
\alias{get_bmaBUGScode}
\title{Extract BMA BUGS code}
\usage{
get_bmaBUGScode(
  option = "smooth",
  print.screen = FALSE,
  save.local = FALSE,
  incl.2deriv = FALSE,
  Y1perfect = FALSE,
  seFromData = FALSE
)
}
\arguments{
\item{option}{text string defining which particular variant is desired}

\item{print.screen}{Logical, should the code be printed to the console?}

\item{save.local}{Logical, should text file of the BUGS code be produced?}

\item{incl.2deriv}{Logical, Should there be a calculation of the indicator's second derivative?}

\item{Y1perfect}{Logical, should the first year for each species be assumed to be known without error (`TRUE`)}

\item{seFromData}{Logical, Should the standard errors be read in from data (`TRUE`) or estimated (`FALSE`)}
}
\description{
Gets a copy of the BUGS code and writes it to the screen and (optionally) to a text file
}
\details{
There are a number of model to choose from:
\itemize{
 \item{\code{"smooth"}}{The default option. Indicator defined by Growth rates, with Ruppert smoother, allowing for species to join late. Error on the first year of each species' time-series is assumed to be zero. The indicator is the expected value of the geometric mean across species (with missing species imputed). 
 Includes three options: `seFromData` `Y1perfect` and `incl.2deriv`. See bayesian_meta_analysis for mode details.}
 \item{\code{"other"}}{XXXXX}
 }
}
\examples{
get_bmaBUGScode(option="smooth", print.screen=TRUE, save.local=FALSE)
get_bmaBUGScode(option="smooth_JABES", print.screen=TRUE, save.local=FALSE)
get_bmaBUGScode(option="smooth_det", print.screen=TRUE, save.local=FALSE)
get_bmaBUGScode(option="smooth_det2", print.screen=TRUE, save.local=FALSE)
get_bmaBUGScode(option="smooth_det_sigtheta", print.screen=TRUE, save.local=FALSE)
}
