% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bmaBUGScode.R
\name{get_bmaBUGScode}
\alias{get_bmaBUGScode}
\title{Extract BMA BUGS code}
\usage{
get_bmaBUGScode(option = "smooth", print.screen = FALSE,
  save.local = FALSE)
}
\arguments{
\item{option}{text string defining which particular variant is desired}

\item{print.screen}{Logical, should the code be printed to the console?}

\item{save.local}{Logical, should text file of the BUGS code be produced?}
}
\description{
Gets a copy of the BUGS code and writes it
}
\examples{
get_bmaBUGScode(option="smooth", print.screen=TRUE, save.local=FALSE)
}
