% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.R
\name{msi}
\alias{msi}
\title{Multi-Species Indicator}
\usage{
msi(data, ...)
}
\arguments{
\item{data}{a data.frame with 4 columns in this order: 'species', 'year', 'index', 'se' (standard error)}

\item{...}{other parameters to pass to \code{msi_tool}}
}
\value{
Returns a dataframe with 4 columns: Year, Index, lower2.5, upper97.5. The last two columns are the credible intervals
}
\description{
A simple wrapper for \code{msi_tool} to make it easier to use in R.
Multi-Species Indicators (MSI) are biodiversity indicators that combine the population
development of species into a single indicator. The MSI-tool calculates an MSI, confidence intervals
for the MSIs and linear and flexible (smoothed) trends. The trends are classified in terms like
"moderate increase", "strong decrease" or "stable". A number of additional analyses can be performed
like testing for changepoints, comparison of trends before and after a changepoint and the calculation
and testing of the total change in a time series.
}
\examples{

# Create some example data in the format required
data <- data.frame(species = rep(letters, each = 50),
                   year = rep(1:50, length(letters)),
                   index = runif(n = 50 * length(letters), min = 50, max = 150),
                   se = runif(n = 50 * length(letters), min = 0.1, max = 8))

# Run the MSI function
msi_out <- msi(data, plot = FALSE)

# Plot the resulting indicator
plot(msi_out)
}
