% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indicator.R
\name{plot_indicator}
\alias{plot_indicator}
\title{Plot indicator}
\usage{
plot_indicator(
  indicator,
  CIs,
  year = 1:length(indicator),
  index = indicator[1],
  smoothed_line = NULL,
  main = ""
)
}
\arguments{
\item{indicator}{A numeric vector, one value for each year}

\item{CIs}{A matrix with two columns, the first column is the lower bound
and the second is the upper bound.}

\item{year}{A numeric vector as long as \code{indicator} giving the years}

\item{index}{Numeric, the index value, by default it is taken to be the 
first value of \code{indicator}}

\item{smoothed_line}{A numeric vector giving the smoothed values to plot
see \code{\link{GAM_smoothing}}}

\item{main}{Title given to the plot}
}
\value{
The plot object is returned
}
\description{
This function plots the indicator using ggplot2.
}
