% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_posterior.r
\name{rescale_posterior}
\alias{rescale_posterior}
\title{Rescale species values for indicator}
\usage{
rescale_posterior(input_dir, subset_table = NULL, index = 100,
  max = 10000, min = 1, year_limit = 10, upperQuantile = 0.975,
  lowerQuantile = 0.025, verbose = TRUE)
}
\arguments{
\item{input_dir}{the location of occupancy model output files}

\item{subset_table}{dataframe with columns for years and rows for species. 
Column names must be character years and row names must match the file names
in input_dir}

\item{index}{The index value for the first year, defaults to 100.}

\item{max}{The upper limit allowed for scaled values. Values greater than 
this will be set equal to \code{max}.}

\item{min}{The upper limit allowed for scaled values. Values greater than 
this will be set equal to \code{min}.}

\item{year_limit}{The minimum length of 'good years' for a species to be
included}

\item{upperQuantile}{The upper confidence interval to use (as a probability)}

\item{lowerQuantile}{The lower confidence interval to use (as a probability)}

\item{verbose}{If TRUE progress is written to console}
}
\value{
A list with two elements, a summary (data.frame) and the rescaled data
(list)
}
\description{
This function takes in the output from a sparta occupancy model and rescales
the posteriers so that the gemean for each species is the same in the first
year. This function accounts for species that have no data at the beginning
of the period, by bringing them in at the geomean, and those that have no
data at the end of the period, by applying a multiplier to the result of the
species data.
}
\details{
There are two options for calculating confidence intervals; the first is
to use the 95% quantiles, this shows the range of values in the posterior, the second
is to calculate the indicator line for each iteration (given in the posterior data)
and then to present the 97.5 and 2.5 quantiles of these lines. Both are returned
in the summary table.
}

