% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rescale_species.R
\name{rescale_species}
\alias{rescale_species}
\title{Rescale species values for indicator}
\usage{
rescale_species(Data, index = 100, max = 10000, min = 1)
}
\arguments{
\item{Data}{A matrix, the first column gives the year, subsequent named
columns give the species. Values in the table are the yearly values
to be rescaled.}

\item{index}{The index value for the first year, defaults to 100.}

\item{max}{The upper limit allowed for scaled values. Values greater than
this will be set equal to \code{max}.}

\item{min}{The upper limit allowed for scaled values. Values greater than
this will be set equal to \code{min}.}
}
\value{
A matrix. In each row we have the year, the species and the
        scaled value. There is also an additional column, 'geomean' giving
        the geometric mean for each year.
}
\description{
This function takes in a dataframe of multi-species data and rescales them
so that the value in the starting year is the same. This function accounts
for species that have no data at the beginning of the period or those that
have no data at teh end of the period.
}

