% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trend_assessment.r
\name{trend_assessment}
\alias{trend_assessment}
\title{Assess trends for indicator}
\usage{
trend_assessment(lambda_output, start_year = NULL, end_year = NULL)
}
\arguments{
\item{lambda_output}{An object returned by lambda_interpolation}

\item{start_year}{(Optional) a numeric value, defaults to the first year}

\item{end_year}{(Optional) a numeric value, defaults to the last year}
}
\value{
Returns a list of two elements, a summary of the species and indicator
assessments. A plot of the species assessment is returned to the device.
}
\description{
This assessment covers both the indicator and the constituant species.
An assessment is made of the indicator over the time period given, examining whether
the initial indicator value falls within the credibal inverval of the final year.
Over the same time period the change in each species is assessed and reported.
}
\examples{
### Running from an array ####
set.seed(123)
# number of species
nsp = 50

# number of years
nyr = 40

#number of iterations
iter = 3

# Build a random set of data
myArray <- array(rnorm(n = nsp*nyr*iter,
                 mean = 0.5,
                 sd = 0.2),
                 c(nsp, nyr, iter))

# Ensure values are bounded by 0 and 1
myArray[myArray > 1] <- 1
myArray[myArray < 0] <- 0

# Run the lambda_interpolation method on this data
myIndicator <- lambda_interpolation(myArray)

# Plot the trend stack
trend_assessment(myIndicator)
}

